Fix use-after-move, integer division and rounding defects
Three use-after-move bugs, all of which changed behaviour:

- TranscriptionService moved onComplete into the `fail` lambda's capture,
  then moved it again into the shared Completion. Every success path got an
  empty std::function, so transcription completion never fired. The shared
  Completion is now built first and `fail` draws on it.

- AIChatConsoleContent moved `error` into ConsoleRunOutput, then tested
  error.empty() 53 lines later to decide whether to append the mixing-agent
  caveat. A moved-from string reads empty, so the caveat appeared even when
  the run had errored. Captured as `hadError` before the move.

- ChordPanelContent moved paintData into the container, then tested
  paintData.rows.empty() for the loading-height branch - always true after
  the move, so the prompt height was applied whatever the row count. Reads
  through the container now.

Seven integer divisions feeding float positions truncated sub-pixel offsets
in note, marker, tick and pill centring. Four `static_cast<int>(x + 0.5)`
roundings are now juce::roundToInt, which is also correct for negatives.

+++ magda/daw/ui/views/SessionView.cpp
@@ -626,7 +626,8 @@ class SessionView::BeatBandContainer : public juce::Component {
 
         const int yCentre = getHeight() / 2;
         // Dot sits in the centre of the column.
-        out.dotCentre = {static_cast<float>(cursor + w / 2), static_cast<float>(yCentre)};
+        out.dotCentre = {static_cast<float>(cursor) + static_cast<float>(w) / 2.0f,
+                         static_cast<float>(yCentre)};
 
         // Icons stacked in the right half, derived from band height.
         constexpr int kGap = 2;
